Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/svg/struct/scripted/autofocus-attribute-expected.txt


FAIL <a> should support autofocus assert_equals: expected Element node <a href="#" autofocus="autofocus"></a> but got null
FAIL Renderable element with tabindex should support autofocus assert_equals: expected Element node <path d="M0,0h8v8z" tabindex="0" autofocus="autofocus"></... but got null
FAIL Never-rendered element with tabindex should not support autofocus promise_test: Unhandled rejection with value: object "TypeError: null is not an object (evaluating 'w.document.activeElement.tagName')"
